Add background fill for transparent images
To ensure that transparent images retain their transparency after
processing, the default background color has been changed to transparent
white (#0fff).

However, if a transparent image is being converted to a format which
doesn't support transparency (e.g, png to jpeg), it will now be properly
rendered onto a white background, as opposed to the black background
that was formerly used when discarding the alpha channel.